backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cssselect2==0.9.0
defusedxml==0.7.1
ecdsa==0.19.1
fastapi==0.133.1
fonttools==4.61.1
fpdf2==2.8.6
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
pypdf==6.7.3
pypdfium2==5.5.0
pyphen==0.17.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
pytest==9.0.2
httpx==0.28.1
pytest-asyncio==1.3.0
